add reminder's controller

+++ app/Http/Controllers/Api/EventController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Models\Event;
 use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Resources\EventResource;
 use Illuminate\Http\JsonResponse;
@@ -14,50 +15,51 @@ class EventController extends BaseController
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
-    public function index(): JsonResponse
+    public function index(Request $request): Response
     {
-        $events = Event::all();
+        $events = $this->getEventsUpcomingOrAll($request)
+            ->orderBy('date', 'asc')
+            ->get();
 
         return $this->sendResponse(EventResource::collection($events), 'Events retrieved successfully.');
     }
+
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return Response
      */
-    public function store(Request $request): JsonResponse
+    public function store(Request $request): Response
     {
         $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'title' => 'required',
-            'date' => 'required'
-        ]);
+        $validator = $this->validateEvent($request);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return $this->sendError('Validation Error.', $validator->errors());
         }
 
+        $input['user_id'] = auth()->id();
+
         $event = Event::create($input);
 
-        return $this->sendResponse(new EventResource($event), 'Event created successfully.');
+        return $this->sendResponse(new EventResource($event), 'Event created successfully.' );
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Event $event
+     * @return Response
      */
-    public function show($id): JsonResponse
+    public function show(Event $event): Response
     {
-        $event = Event::find($id);
-
-        if (is_null($event)) {
-            return $this->sendError('Event not found.');
+        if (!$this->checkUserId($event)) {
+            return $this->sendError('Unauthorized.');
         }
 
         return $this->sendResponse(new EventResource($event), 'Event retrieved successfully.');
@@ -66,25 +68,26 @@ class EventController extends BaseController
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param Event $event
+     * @return Response
      */
-    public function update(Request $request, Event $event): JsonResponse
+    public function update(Request $request, Event $event): Response
     {
+        if (!$this->checkUserId($event)) {
+            return $this->sendError('Unauthorized.');
+        }
+
         $input = $request->all();
 
-        $validator = Validator::make($input, [
-            'name' => 'required',
-            'date' => 'required'
-        ]);
+        $validator = $this->validateEvent($request);
 
-        if($validator->fails()){
+        if ($validator->fails()) {
             return $this->sendError('Validation Error.', $validator->errors());
         }
 
-        $event->name = $input['name'];
-        $event->detail = $input['date'];
+        $event->title = $input['title'];
+        $event->date = $input['date'];
         $event->save();
 
         return $this->sendResponse(new EventResource($event), 'Event updated successfully.');
@@ -93,13 +96,39 @@ class EventController extends BaseController
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Event $event
+     * @return Response
      */
-    public function destroy(Event $event): JsonResponse
+    public function destroy(Event $event): Response
     {
+        if (!$this->checkUserId($event)) {
+            return $this->sendError('Unauthorized.');
+        }
+
         $event->delete();
 
         return $this->sendResponse([], 'Event deleted successfully.');
     }
+
+    private function validateEvent(Request $request): \Illuminate\Validation\Validator
+    {
+        return Validator::make($request->all(), [
+            'title' => 'required',
+            'date' => 'required|date|after_or_equal:' . now(),
+        ]);
+    }
+
+    private function checkUserId(Event $event): bool
+    {
+        return $event->user_id == auth()->id();
+    }
+
+    private function getEventsUpcomingOrAll(Request $request): \Illuminate\Database\Eloquent\Builder
+    {
+        if ($request->query('past', 'false') === 'true') return Event::where('user_id', auth()->id());
+
+        return Event::where('user_id', auth()->id())
+            ->where('date', '>=', now());
+    }
 }
+
